Fully separated single exercise page and exercises into separate paths
