+++ src/Pages/Exercises/Exercises.tsx
@@ -1,5 +1,5 @@
 import "./Exercises.scss";
-import { Await, Outlet, useLoaderData, useNavigate } from "react-router-dom";
+import { Await, useLoaderData, useNavigate } from "react-router-dom";
 import { Suspense } from "react";
 import { Immutable, Narrow } from "../../Types/Utility/Models";
 import { FullExercise } from "../../Types/Models/FullExercise";
@@ -16,22 +16,18 @@ export default function Exercises() {
           exercises: Immutable<Narrow<FullExercise, ["id", "name", "image"]>>[]
         ) => (
           <div className="exercises-page-container">
-            <Outlet />
-
-            <div className="exercise-cards-container">
-              {exercises.map((exercise) => {
-                return (
-                  <div
-                    className="exercise-card"
-                    key={exercise.id}
-                    onClick={() => navigate(`/exercises/${exercise.id}`)}
-                  >
-                    <p>{exercise.name}</p>
-                    <img src={exercise.image} alt="" />
-                  </div>
-                );
-              })}
-            </div>
+            {exercises.map((exercise) => {
+              return (
+                <div
+                  className="exercise-card"
+                  key={exercise.id}
+                  onClick={() => navigate(`/exercises/${exercise.id}`)}
+                >
+                  <p>{exercise.name}</p>
+                  <img src={exercise.image} alt="" />
+                </div>
+              );
+            })}
           </div>
         )}
       </Await>
