+++ src/main.tsx
@@ -2,17 +2,10 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.scss";
-import {
-  Outlet,
-  RouterProvider,
-  createBrowserRouter,
-  defer,
-} from "react-router-dom";
+import { RouterProvider, createBrowserRouter, defer } from "react-router-dom";
 import Error from "./Components/Error/Error.tsx";
 import Exercises from "./Pages/Exercises/Exercises.tsx";
-import FullExerciseDisplay, {
-  singleExerciseLoader,
-} from "./Components/FullExerciseDisplay/FullExerciseDisplay.tsx";
+import FullExerciseDisplay from "./Components/FullExerciseDisplay/FullExerciseDisplay.tsx";
 import Profile from "./Pages/Profile/Profile.tsx";
 import Authentication from "./Pages/Authentication/Authentication.tsx";
 import { getCurrentUserData } from "./Data/User.ts";
@@ -21,6 +14,7 @@ import AdminExercisePanel from "./Pages/AdminPanel/Exercises/AdminExercisePanel.
 import NewExercise from "./Pages/AdminPanel/Exercises/New/NewExercise.tsx";
 import { newExerciseLoader } from "./Pages/AdminPanel/Exercises/New/NewExerciseLoader.tsx";
 import exerciseLoader from "./Pages/Exercises/ExerciseLoader.ts";
+import singleExerciseLoader from "./Components/FullExerciseDisplay/SingleExerciseLoader.ts";
 
 const router = createBrowserRouter([
   {
@@ -44,18 +38,11 @@ const router = createBrowserRouter([
       },
       {
         path: "workouts",
-        element: (
-          <div>
-            Workouts
-            <Outlet />
-          </div>
-        ),
-        children: [
-          {
-            path: ":workoutId",
-            element: <div>workout</div>,
-          },
-        ],
+        element: <div>Workouts</div>,
+      },
+      {
+        path: "workouts/:workoutId",
+        element: <div>workout</div>,
       },
       {
         path: "me",
